+++ fixtures/app/home/scripts/build.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+echo "Building your home"
+sleep 2
+echo "Processing artifacts"
+sleep 2
+echo "Transpiling Javascript"
+sleep 2