+++ packages/app/src/cli/models/app/app.ts
@@ -1,8 +1,12 @@
 import {file, error, path, schema, string, toml} from '@shopify/cli-kit'
-import {blocks, configurationFileNames, genericConfigurationFileNames, uiExtensions} from '$cli/constants'
+import {blocks, configurationFileNames, genericConfigurationFileNames, extensions} from '$cli/constants'
 
 export const HomeNotFoundError = (homeDirectory: string) => {
   return new error.Abort(`Couldn't find the home directory at ${homeDirectory}`)
+}
+
+export const HomeConfigurationFileNotFound = (configurationFilePath: string) => {
+  return new error.Abort(`Couldn't find the home configuraiton file at ${configurationFilePath}`)
 }
 
 export const AppConfigurationSchema = schema.define.object({
@@ -12,12 +16,12 @@ export const AppConfigurationSchema = schema.define.object({
 
 type AppConfiguration = schema.define.infer<typeof AppConfigurationSchema>
 
-const UIExtensionConfigurationSchema = schema.define.object({
+const ExtensionConfigurationSchema = schema.define.object({
   name: schema.define.string(),
-  type: schema.define.enum(uiExtensions.types),
+  type: schema.define.enum(extensions.types),
 })
 
-type UIExtensionConfiguration = schema.define.infer<typeof UIExtensionConfigurationSchema>
+type ExtensionConfiguration = schema.define.infer<typeof ExtensionConfigurationSchema>
 
 const ScriptConfigurationSchema = schema.define.object({
   name: schema.define.string(),
@@ -30,13 +34,24 @@ interface Script {
   directory: string
 }
 
-interface UIExtension {
-  configuration: UIExtensionConfiguration
+export interface Extension {
+  configuration: ExtensionConfiguration
   directory: string
+  buildDirectory: string
 }
 
-interface Home {
+export const HomeConfigurationSchema = schema.define.object({
+  commands: schema.define.object({
+    build: schema.define.string().optional(),
+    dev: schema.define.string(),
+  }),
+})
+
+export type HomeConfiguration = schema.define.infer<typeof HomeConfigurationSchema>
+export type HomeConfigurationCommands = keyof HomeConfiguration['commands']
+export interface Home {
   directory: string
+  configuration: HomeConfiguration
 }
 
 type PackageManager = 'npm' | 'yarn' | 'pnpm'
@@ -47,7 +62,7 @@ export interface App {
   configuration: AppConfiguration
   scripts: Script[]
   home: Home
-  uiExtensions: UIExtension[]
+  extensions: Extension[]
 }
 
 export async function load(directory: string): Promise<App> {
@@ -64,7 +79,7 @@ export async function load(directory: string): Promise<App> {
   const configuration = await parseConfigurationFile(AppConfigurationSchema, configurationPath)
   const appDirectory = path.dirname(configurationPath)
   const scripts = await loadScripts(appDirectory)
-  const uiExtensions = await loadUiExtensions(appDirectory)
+  const extensions = await loadExtensions(appDirectory)
   const yarnLockPath = path.join(appDirectory, genericConfigurationFileNames.yarn.lockfile)
   const yarnLockExists = await file.exists(yarnLockPath)
   const pnpmLockPath = path.join(appDirectory, genericConfigurationFileNames.pnpm.lockfile)
@@ -84,7 +99,7 @@ export async function load(directory: string): Promise<App> {
     configuration,
     home,
     scripts,
-    uiExtensions,
+    extensions,
     packageManager,
   }
 }
@@ -94,7 +109,12 @@ async function loadHome(appDirectory: string): Promise<Home> {
   if (!(await file.exists(homeDirectory))) {
     throw HomeNotFoundError(homeDirectory)
   }
-  return {directory: homeDirectory}
+  const homeConfigurationFile = path.join(homeDirectory, configurationFileNames.home)
+  if (!(await file.exists(homeConfigurationFile))) {
+    throw HomeConfigurationFileNotFound(homeConfigurationFile)
+  }
+  const configuration = await parseConfigurationFile(HomeConfigurationSchema, homeConfigurationFile)
+  return {directory: homeDirectory, configuration}
 }
 
 async function loadConfigurationFile(path: string): Promise<object> {
@@ -117,16 +137,16 @@ async function parseConfigurationFile(schema: any, path: string) {
   return parseResult.data
 }
 
-async function loadUiExtensions(rootDirectory: string): Promise<UIExtension[]> {
-  const uiExtensionsPath = path.join(rootDirectory, `${blocks.uiExtensions.directoryName}/*`)
-  const directories = await path.glob(uiExtensionsPath, {onlyDirectories: true})
-  return Promise.all(directories.map((directory) => loadUiExtension(directory)))
+async function loadExtensions(rootDirectory: string): Promise<Extension[]> {
+  const extensionsPath = path.join(rootDirectory, `${blocks.extensions.directoryName}/*`)
+  const directories = await path.glob(extensionsPath, {onlyDirectories: true})
+  return Promise.all(directories.map((directory) => loadExtension(directory)))
 }
 
-async function loadUiExtension(directory: string): Promise<UIExtension> {
-  const configurationPath = path.join(directory, blocks.uiExtensions.configurationName)
-  const configuration = await parseConfigurationFile(UIExtensionConfigurationSchema, configurationPath)
-  return {directory, configuration}
+async function loadExtension(directory: string): Promise<Extension> {
+  const configurationPath = path.join(directory, blocks.extensions.configurationName)
+  const configuration = await parseConfigurationFile(ExtensionConfigurationSchema, configurationPath)
+  return {directory, configuration, buildDirectory: path.join(directory, 'build')}
 }
 
 async function loadScripts(rootDirectory: string): Promise<Script[]> {