+++ packages/create-app/src/prompts/init.test.ts
@@ -4,8 +4,8 @@ import {describe, it, expect, vi} from 'vitest'
 describe('init', () => {
   it('when name is not passed', async () => {
     const prompt = vi.fn()
-    const answers = {name: 'app'}
-    const options = {}
+    const answers = {name: 'app', template: 'https://github.com/Shopify/shopify-app-node#cli-next'}
+    const options = {template: 'template'}
 
     // Given
     prompt.mockResolvedValue(Promise.resolve(answers))
@@ -28,8 +28,8 @@ describe('init', () => {
 
   it('when name is passed', async () => {
     const prompt = vi.fn()
-    const answers = {name: 'app'}
-    const options = {name: 'app'}
+    const answers = {name: 'app', template: 'https://github.com/Shopify/shopify-app-node#cli-next'}
+    const options = {name: 'app', template: 'template'}
 
     // Given
     prompt.mockResolvedValue(Promise.resolve(answers))