+++ packages/create-hydrogen/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @shopify/create-hydrogen
 
+## 1.0.1
+
+### Patch Changes
+
+- Some fixes
+- Updated dependencies
+  - @shopify/cli-hydrogen@1.0.1
+  - @shopify/cli-kit@1.0.1
+
+## 1.0.0
+
+### Patch Changes
+
+- Get workflows working e2e
+- Updated dependencies
+  - @shopify/cli-kit@1.0.0
+  - @shopify/cli-hydrogen@1.0.0
+
+## 0.33.8
+
+### Patch Changes
+
+- Remove unnecessary dependencies
+- Updated dependencies
+  - @shopify/cli-hydrogen@0.33.8
+  - @shopify/cli-kit@0.33.8
+
+## 0.33.7
+
+### Patch Changes
+
+- Add .gitignore to the app template
+- Updated dependencies
+  - @shopify/cli-hydrogen@0.33.7
+  - @shopify/cli-kit@0.33.7
+
+## 0.33.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@0.33.6
+
+## 0.33.5
+
+### Patch Changes
+
+- Some fixes with external packages
+- Updated dependencies
+  - @shopify/cli-kit@0.33.5
+
+## 0.33.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@0.33.3
+  - @shopify/cli-hydrogen@0.33.3
+
 ## 0.33.2
 
 ### Patch Changes